Add pixels module
