+++ doc/old/working_with_pixel_data.rst
@@ -127,7 +127,7 @@ of ``3`` then the returned pixel data will be in the color space as given by
 
 *pydicom* offers a limited ability to convert between 8-bits/channel YBR and
 RGB color spaces through the
-:func:`~pydicom.pixel_data_handlers.util.convert_color_space`
+:func:`~pydicom.pixels.utils.convert_color_space`
 function. When changing the color space you should also change the value
 of *Photometric Interpretation* to match.
 
@@ -144,12 +144,12 @@ Some DICOM datasets store their output image pixel values in a lookup table
 (LUT), where the values in *Pixel Data* are the index to a corresponding
 LUT entry. When a dataset's (0028,0004) *Photometric Interpretation* value is
 ``PALETTE COLOR`` then the
-:func:`~pydicom.pixel_data_handlers.util.apply_color_lut` function can be used
+:func:`~pydicom.pixels.utils.apply_color_lut` function can be used
 to apply a palette color LUT to the pixel data to produce an RGB image.
 
 .. code-block:: python
 
-    from pydicom.pixel_data_handlers.util import apply_color_lut
+    from pydicom.pixels import apply_color_lut
 
     fname = get_testdata_file("OBXXXX1A.dcm")
     ds = dcmread(fname)
@@ -163,7 +163,7 @@ of the pixel data is 8-bit.
 
 .. code-block:: python
 
-    from pydicom.pixel_data_handlers.util import apply_color_lut
+    from pydicom.pixels import apply_color_lut
 
     fname = get_testdata_file("OBXXXX1A.dcm")
     ds = dcmread(fname)
@@ -185,7 +185,7 @@ Modality LUT or Rescale Operation
 The DICOM :dcm:`Modality LUT<part03/sect_C.11.html#sect_C.11.1>` module
 converts raw pixel data values to a specific (possibly unitless) physical
 quantity, such as Hounsfield units for CT. The
-:func:`~pydicom.pixel_data_handlers.util.apply_modality_lut` function can be
+:func:`~pydicom.pixels.utils.apply_modality_lut` function can be
 used with an input array of raw values and a dataset containing a Modality LUT
 module to return the converted values. When a dataset requires multiple
 grayscale transformations, the Modality LUT transformation is always applied
@@ -193,7 +193,7 @@ first.
 
 .. code-block:: python
 
-    from pydicom.pixel_data_handlers.util import apply_modality_lut
+    from pydicom.pixels import apply_modality_lut
 
     fname = get_testdata_file("CT_small.dcm")
     ds = dcmread(fname)
@@ -206,7 +206,7 @@ VOI LUT or Windowing Operation
 
 The DICOM :dcm:`VOI LUT<part03/sect_C.11.2.html>` module applies a
 VOI or windowing operation to input values. The
-:func:`~pydicom.pixel_data_handlers.util.apply_voi_lut` function
+:func:`~pydicom.pixels.utils.apply_voi_lut` function
 can be used with an input array and a dataset containing a VOI LUT module to
 return values with applied VOI LUT or windowing. When a dataset contains
 multiple VOI or windowing views then a particular view can be returned by
@@ -217,7 +217,7 @@ that the modality LUT or rescale operation has already been applied.
 
 .. code-block:: python
 
-    from pydicom.pixel_data_handlers.util import apply_voi_lut
+    from pydicom.pixels import apply_voi_lut
 
     fname = get_testdata_file("MR-SIEMENS-DICOM-WithOverlays.dcm")
     ds = dcmread(fname)
