+++ doc/release_notes/v2.0.0.rst
@@ -51,8 +51,8 @@ Fixes
 * Fixed error when writing elements with a VR of **OF** (:issue:`1075`)
 * Fixed improper conversion when reading elements with a VR of **OF**
   (:issue:`1075`)
-* Fixed :func:`~pydicom.pixel_data_handlers.util.apply_voi_lut` and
-  :func:`~pydicom.pixel_data_handlers.util.apply_modality_lut` not handling
+* Fixed :func:`~pydicom.pixels.utils.apply_voi_lut` and
+  :func:`~pydicom.pixels.utils.apply_modality_lut` not handling
   (0028,3006) *LUT Data* with a VR of **OW** (:issue:`1073`)
 * Fixed access to private creator tag in raw datasets (:issue:`1078`)
 * Fixed description of newly added known private tag (:issue:`1082`)
