+++ doc/release_notes/v2.2.0.rst
@@ -23,7 +23,7 @@ Changes
 * :func:`~pydicom.pixel_data_handlers.rle_handler.rle_encode_frame` is
   deprecated and will be removed in v3.0, use
   :meth:`~pydicom.dataset.Dataset.compress` or
-  :attr:`~pydicom.encoders.RLELosslessEncoder` instead.
+  :attr:`~pydicom.pixels.encoders.RLELosslessEncoder` instead.
 * :func:`~pydicom.filereader.read_file` is deprecated and will be removed in
   v3.0, use :func:`~pydicom.filereader.dcmread` instead.
 * :func:`~pydicom.filewriter.write_file` is deprecated and will be removed in
@@ -72,7 +72,7 @@ Enhancements
 * Added support for downloading the large test files with the `requests
   <https://docs.python-requests.org/en/master/>`_ package in addition to
   :mod:`urllib.request` (:pr:`1340`)
-* Ensured :func:`~pydicom.pixel_data_handlers.util.convert_color_space` uses
+* Ensured :func:`~pydicom.pixels.utils.convert_color_space` uses
   32-bit floats for calculation, added `per_frame` flag to allow frame-by-frame
   processing and improved the speed by ~20-60% (:issue:`1348`)
 * Optimisations for RLE encoding using *pydicom* (~40% faster).
@@ -82,9 +82,9 @@ Enhancements
 * Added :func:`Dataset.compress()<pydicom.dataset.Dataset.compress>` function for
   compressing uncompressed pixel data using a given encoding format as specified
   by a UID. Only *RLE Lossless* is currently supported (:pr:`1372`)
-* Added :mod:`~pydicom.encoders` module and the following encoders:
+* Added :mod:`~pydicom.pixels.encoders` module and the following encoders:
 
-  * :attr:`~pydicom.encoders.RLELosslessEncoder` with 'pydicom', 'pylibjpeg'
+  * :attr:`~pydicom.pixels.encoders.RLELosslessEncoder` with 'pydicom', 'pylibjpeg'
     and 'gdcm' plugins
 * Added `read` parameter to :func:`~pydicom.data.get_testdata_file`
   to allow reading and returning the corresponding dataset (:pr:`1372`)
