+++ doc/release_notes/v2.3.0.rst
@@ -9,6 +9,23 @@ Changes
 * Updated the private dictionary
 * :attr:`~pydicom.config.enforce_valid_values` is deprecated in favor of
   :attr:`~pydicom.config.settings.reading_validation_mode`
+* The ``pydicom.pixel_data_handlers`` module is deprecated and will be removed in
+  v4.0. The following functions should be imported from the new ``pydicom.pixels``
+  module instead:
+
+  * :func:`~pydicom.pixels.utils.apply_color_lut`
+  * :func:`~pydicom.pixels.utils.apply_modality_lut`
+  * :func:`~pydicom.pixels.utils.apply_rescale`
+  * :func:`~pydicom.pixels.utils.apply_voi`
+  * :func:`~pydicom.pixels.utils.apply_voi_lut`
+  * :func:`~pydicom.pixels.utils.apply_windowing`
+  * :func:`~pydicom.pixels.utils.convert_color_space`
+  * :func:`~pydicom.pixels.utils.pack_bits`
+  * :func:`~pydicom.pixels.utils.unpack_bits`
+
+* The ``pydicom.encoders`` module has been moved to ``pydicom.pixels.encoders``,
+  the old import path will be removed in v3.0
+
 
 
 Enhancements
@@ -25,7 +42,7 @@ Enhancements
   the slowest handler (:issue:`1487`)
 * Added, enhanced, or removed a number of Mitra private dictionary entries (:pr:`1588`)
 * Added support for unpacking bit-packed data without using NumPy to
-  :func:`~pydicom.pixel_data_handlers.utils.unpack_bits`(:pr:`1594`)
+  :func:`~pydicom.pixels.utils.unpack_bits`(:pr:`1594`)
 
 
 Fixes
