+++ pydicom/dataset.py
@@ -51,7 +51,7 @@ from pydicom.datadict import (
 from pydicom.dataelem import DataElement, DataElement_from_raw, RawDataElement
 from pydicom.encaps import encapsulate, encapsulate_extended
 from pydicom.fileutil import path_from_pathlike, PathType
-from pydicom.pixel_data_handlers.util import (
+from pydicom.pixels.utils import (
     convert_color_space, reshape_pixel_array, get_image_pixel_ids
 )
 from pydicom.tag import Tag, BaseTag, tag_in_exception, TagType
@@ -1668,7 +1668,7 @@ class Dataset:
             present. See the :doc:`encoding plugins options
             </guides/encoding/encoder_plugin_options>` for more information.
         """
-        from pydicom.encoders import get_encoder
+        from pydicom.pixels import get_encoder
 
         uid = UID(transfer_syntax_uid)
