+++ pydicom/pixels/__init__.py
@@ -1,0 +1,13 @@
+from pydicom.pixels.encoders.base import get_encoder, RLELosslessEncoder
+from pydicom.pixels.utils import (
+    apply_color_lut,
+    apply_modality_lut,
+    apply_voi_lut,
+    convert_color_space,
+    apply_voi,
+    apply_windowing,
+    pack_bits,
+    unpack_bits,
+)
+
+apply_rescale = apply_modality_lut
