+++ pydicom/pixels/encoders/gdcm.py
