+++ pydicom/pixels/encoders/pylibjpeg.py
