+++ pydicom/tests/pixels/test_handlers_rle.py
@@ -30,12 +30,12 @@ from pydicom import dcmread
 import pydicom.config
 from pydicom.data import get_testdata_file
 from pydicom.encaps import defragment_data
+from pydicom.pixels.utils import reshape_pixel_array
 from pydicom.uid import RLELossless, AllTransferSyntaxes
 
 try:
     import numpy as np
     from pydicom.pixel_data_handlers import numpy_handler as NP_HANDLER
-    from pydicom.pixel_data_handlers.util import reshape_pixel_array
     HAVE_NP = NP_HANDLER.HAVE_NP
 except ImportError:
     NP_HANDLER = None
